+++ src/api/bankApi.js
@@ -1,31 +1,26 @@
 // axios 라이브러리 불러오기 -> 커스텀 인스턴스 만들기
 import axiosInstance from './axios';
 
-// Codef - 계좌 연동 & 목록 불러오기
-export const fetchAccountsFromCodef = async ({
-  bankId,
-  userBankId,
-  userBankPassword,
-}) => {
+// 계좌 목록 조회
+export const fetchAccounts = async () => {
   try {
-    const { data } = await axiosInstance.post('/api/codef/account/add', {
-      bankId,
-      userBankId,
-      userBankPassword,
-    });
-    return data;
-  } catch (e) {
-    console.error('계좌 연동 실패: ', e);
-    throw new Error('계좌 연동 실패');
+    const response = await axiosInstance.get('/account/db');
+    return response.data;
+  } catch (error) {
+    console.error('계좌 목록 불러오기 실패:', error);
+    throw error; // 필요 시 상위 컴포넌트에서 catch 처리
   }
 };
 
-export const getAccountList = async () => {
+// 거래 내역 조회
+export const fetchTransactions = async accountNum => {
   try {
-    const { data } = await axiosInstance.get('/account/db');
+    const { data } = await axiosInstance.get(
+      `/transaction/db?account=${accountNum}`
+    );
     return data;
   } catch (e) {
-    console.error('계좌 목록 조회 실패: ', e);
-    throw new Error('계좌 목록 조회 실패');
+    console.error('거래내역 조회 실패: ', e);
+    throw new Error('거래내역 조회 실패');
   }
 };
