+++ src/views/home/HomeView.vue
@@ -116,7 +116,7 @@
 import { onMounted, ref } from 'vue';
 import { useRouter } from 'vue-router';
 
-import { fetchAccounts } from '@/api/accountApi';
+import { fetchAccounts } from '@/api/bankApi';
 import img_character from '@/assets/img/characters/character_zero_1.png';
 import icon_plus from '@/assets/img/icons/feature/icon_plus.png';
 import icon_refresh from '@/assets/img/icons/feature/icon_refresh.png';
